requirements.txt
einops==0.8.2
loguru==0.7.3
matplotlib==3.10.8
numpy==2.4.1
Pillow==12.1.0
rich==14.3.1
scikit_learn==1.8.0
timm==1.0.24
torch>=2.0.0
torchvision>=0.15.0
tqdm==4.67.1
wandb>=0.15.0
